Add weather_station client and server
The added files simulate a ZeroMQ Publisher and Subscriber connection
over tcp socket. The Publisher socket is bound to the server where
weather information is constantly published to. The Subscriber socket is
connected via the client with a specific postal code topic, which
filters out a subset of all published messages matching the topic. The
client averages 10 of these updates and prints the result.